move calculation to a hook
